rev week4

+++ database/migrations/2021_08_23_064419_create_sell_summaries_table.php
@@ -20,7 +20,7 @@ class CreateSellSummariesTable extends Migration
             $table->dateTime('created_date');
             $table->dateTime('last_update');
             $table->integer('price_total');
-            $table->decimal('discount_total');
+            $table->integer('discount_total');
             $table->integer('total');
             $table->timestamps();
         });
